{{cookiecutter.project_slug}}/flask-image-service/requirements.txt
Flask
zappa
boto3
Pillow
